+++ src/funcs/deploymentsCreateDeployment.ts
@@ -16,16 +16,16 @@ import {
   CreateDeploymentResponseBody$inboundSchema,
 } from "../models/createdeploymentresponsebody.js";
 import {
-  CreateDeploymentRequest,
-  CreateDeploymentRequest$outboundSchema,
-} from "../models/createdeploymentresponsebodydeploymentsstatus.js";
-import {
   ConnectionError,
   InvalidRequestError,
   RequestAbortedError,
   RequestTimeoutError,
   UnexpectedClientError,
 } from "../models/httpclienterrors.js";
+import {
+  CreateDeploymentRequest,
+  CreateDeploymentRequest$outboundSchema,
+} from "../models/responsebodyimages.js";
 import { ResponseValidationError } from "../models/responsevalidationerror.js";
 import { SDKValidationError } from "../models/sdkvalidationerror.js";
 import { VercelError } from "../models/vercelerror.js";