+++ RELEASES.md
@@ -1809,3 +1809,13 @@ Based on:
 - [typescript v1.28.34] .
 ### Releases
 - [NPM v1.28.34] https://www.npmjs.com/package/@vercel/sdk/v/1.28.34 - .
+
+## 2026-09-17 00:11:10
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.797.0 (2.937.18) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v1.28.35] .
+### Releases
+- [NPM v1.28.35] https://www.npmjs.com/package/@vercel/sdk/v/1.28.35 - .